fixed jparepository extensions on quotes and quotesproducts daos

+++ src/main/java/com/bpatel2001/invoicing_system/dao/QuoteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(path="quotes")
-public interface QuoteRepository extends JpaRepository<Quotes, Integer> {
+public interface QuoteRepository extends JpaRepository<Quotes, Long> {
 }
